+++ server/src/main/java/org/opensearch/rest/action/cat/RestSegmentsAction.java
@@ -44,7 +44,10 @@ import org.opensearch.cluster.node.DiscoveryNodes;
 import org.opensearch.common.Table;
 import org.opensearch.common.logging.DeprecationLogger;
 import org.opensearch.core.common.Strings;
+import org.opensearch.core.common.breaker.CircuitBreaker;
+import org.opensearch.core.common.breaker.CircuitBreakingException;
 import org.opensearch.index.engine.Segment;
+import org.opensearch.rest.RequestLimitSettings;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.rest.RestResponse;
 import org.opensearch.rest.action.RestActionListener;
@@ -55,6 +58,7 @@ import java.util.Map;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.unmodifiableList;
+import static org.opensearch.rest.RequestLimitSettings.BlockAction.CAT_SEGMENTS;
 import static org.opensearch.rest.RestRequest.Method.GET;
 
 /**
@@ -65,6 +69,12 @@ import static org.opensearch.rest.RestRequest.Method.GET;
 public class RestSegmentsAction extends AbstractCatAction {
 
     private static final DeprecationLogger deprecationLogger = DeprecationLogger.getLogger(RestSegmentsAction.class);
+
+    private final RequestLimitSettings requestLimitSettings;
+
+    public RestSegmentsAction(RequestLimitSettings requestLimitSettings) {
+        this.requestLimitSettings = requestLimitSettings;
+    }
 
     @Override
     public List<Route> routes() {
@@ -78,6 +88,11 @@ public class RestSegmentsAction extends AbstractCatAction {
 
     @Override
     public boolean allowSystemIndexAccessByDefault() {
+        return true;
+    }
+
+    @Override
+    public boolean isRequestLimitCheckSupported() {
         return true;
     }
 
@@ -96,6 +111,10 @@ public class RestSegmentsAction extends AbstractCatAction {
         return channel -> client.admin().cluster().state(clusterStateRequest, new RestActionListener<ClusterStateResponse>(channel) {
             @Override
             public void processResponse(final ClusterStateResponse clusterStateResponse) {
+                if (isRequestLimitCheckSupported()
+                    && requestLimitSettings.isCircuitLimitBreached(clusterStateResponse.getState(), CAT_SEGMENTS)) {
+                    throw new CircuitBreakingException("Segments from too many indices requested.", CircuitBreaker.Durability.TRANSIENT);
+                }
                 final IndicesSegmentsRequest indicesSegmentsRequest = new IndicesSegmentsRequest();
                 indicesSegmentsRequest.indices(indices);
                 client.admin().indices().segments(indicesSegmentsRequest, new RestResponseListener<IndicesSegmentResponse>(channel) {