+++ server/src/test/java/org/opensearch/rest/action/cat/RestIndicesActionTests.java
@@ -43,10 +43,12 @@ import org.opensearch.cluster.routing.ShardRoutingState;
 import org.opensearch.cluster.routing.TestShardRouting;
 import org.opensearch.common.Table;
 import org.opensearch.common.UUIDs;
+import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.core.index.Index;
 import org.opensearch.core.index.shard.ShardId;
 import org.opensearch.index.IndexSettings;
+import org.opensearch.rest.RequestLimitSettings;
 import org.opensearch.test.OpenSearchTestCase;
 import org.opensearch.test.rest.FakeRestRequest;
 
@@ -137,7 +139,10 @@ public class RestIndicesActionTests extends OpenSearchTestCase {
             }
         }
 
-        final RestIndicesAction action = new RestIndicesAction();
+        final ClusterSettings clusterSettings = new ClusterSettings(Settings.EMPTY, ClusterSettings.BUILT_IN_CLUSTER_SETTINGS);
+        final Settings settings = Settings.builder().build();
+        final RequestLimitSettings requestLimitSettings = new RequestLimitSettings(clusterSettings, settings);
+        final RestIndicesAction action = new RestIndicesAction(requestLimitSettings);
         final Table table = action.buildTable(new FakeRestRequest(), indicesSettings, indicesHealths, indicesStats, indicesMetadatas);
 
         // now, verify the table is correct